feat(sys)!: correct types for the integer constants
Stet up bindgen ParseCallbacks with types for common NGX_ constants.
This reduces the need of type casts when passing said constants to ngx_
methods and improves type safety.

The configuration may be incomplete and will be adjusted in the future.

+++ examples/upstream.rs
@@ -265,7 +265,7 @@ unsafe extern "C" fn ngx_http_upstream_commands_set_custom(
 
     if let Some(value) = args.get(1) {
         let n = unsafe { ngx_atoi(value.data, value.len) };
-        if n == (NGX_ERROR as isize) || n == 0 {
+        if n == NGX_ERROR || n == 0 {
             ngx_conf_log_error!(
                 NGX_LOG_EMERG,
                 cf,
